refactor(agent): rename test helper to in-memory-transport
