+++ packages/agent/test/runtime/messenger/history.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, it } from "bun:test";
 import { AgentMessenger } from "../../../src/runtime/messenger/messenger";
 import { queryHistory } from "../../../src/runtime/messenger/history";
 import type { Messenger } from "@openomni/protocol";
-import { InMemoryTransport } from "./helpers";
+import { InMemoryTransport } from "./in-memory-transport";
 
 function makeEnvelope(
   from: string,
